+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: Fixed spurious Z rotations during speherical blend.
 - Bugfix: Blending speed was not set correctly, when blending back and forth between the same cameras.
 - Regression fix: POV is relative to its parent transform.
+- Bugfix: AxisState.Recentering.RecenterNow() did not work reliably.
+
 
 ## [2.8.6] - 2022-05-03
 - Bugfix: A memory leak no longer occurs with PostProcessing if no PP layer is present on the camera.